refactor(frontend): split toast store from component file and sort spiderfy imports
