+++ frontend/src/components/toast.tsx
@@ -1,28 +1,6 @@
 import { useEffect, useState } from "react";
 
-export type ToastKind = "info" | "error" | "success";
-
-interface ToastItem {
-  id: number;
-  message: string;
-  kind: ToastKind;
-  duration: number;
-}
-
-type Listener = (t: ToastItem) => void;
-const listeners = new Set<Listener>();
-let nextId = 1;
-
-/** Fire a transient toast. No-op if no <ToastHost> is mounted. */
-export function toast(message: string, opts?: { kind?: ToastKind; duration?: number }): void {
-  const item: ToastItem = {
-    id: nextId++,
-    message,
-    kind: opts?.kind ?? "info",
-    duration: opts?.duration ?? 4500,
-  };
-  listeners.forEach((l) => l(item));
-}
+import { subscribeToasts, type ToastItem, type ToastKind } from "./toastStore";
 
 const KIND_CLASS: Record<ToastKind, string> = {
   info: "bg-gray-900/85 border-white/10 text-gray-200",
@@ -35,12 +13,10 @@ export function ToastHost() {
   const [items, setItems] = useState<ToastItem[]>([]);
 
   useEffect(() => {
-    const onToast: Listener = (t) => {
+    return subscribeToasts((t) => {
       setItems((prev) => [...prev, t]);
       window.setTimeout(() => setItems((prev) => prev.filter((x) => x.id !== t.id)), t.duration);
-    };
-    listeners.add(onToast);
-    return () => { listeners.delete(onToast); };
+    });
   }, []);
 
   if (items.length === 0) return null;
